Adding a new problem for graduates due 2/16
Homework 1 Problem 7 (optional for 498 students) is now released!  Because it involves running things on the GPU, we are releasing it as a Google Colab notebook (which will also serve as the template for you to turn in to us). This problem is due on Feb 16 for 599 students. ROB498 students are still encouraged to at least read the assignment and the papers referenced within, but completion is not required.

You will find the template here: https://colab.research.google.com/drive/1RZyRTmjMjDpuCpkJKFqrnJdR87s_zqyP?usp=sharing

You should not be able to edit that link - instead, make a copy of the notebook into your own Google Drive, and rename it to match your uniqname (e.g. Homework1_Problem7_Stanlew) to aid in submission/grading.

As a part of this notebook, you will be comparing your results from the main HW1 reconstruction to one created by a deeply learned vision model called Mast3r. Everything should already be setup to run the model in this notebook, but if you run into any issues please don't hesitate to ask (as getting that infrastructure setup is definitely outside the scope of this assignment!).